pyproject.toml
[project]
name = "neoskills"
version = "0.1.0"
description = "Cross-Agent Skill Bank, Transfer System, and Embedded Skill Substrate"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "Richard Tong" }]
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "click>=8.1",
    "gitpython>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]
sdk = [
    "claude-agent-sdk>=0.1.0",
    "anthropic>=0.40.0",
]

[project.scripts]
neoskills = "neoskills.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neoskills"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]
